pyproject.toml
[project]
name = "tyran"
version = "0.7.2"
description = "A Vector Search as a Service"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi[standard]==0.121.0",
    "jsonschema==4.25.1",
    "mysqlclient==2.2.7",
    "openai==2.7.1",
    "pydantic-settings==2.11.0",
    "pytest==8.4.2",
    "python-dotenv==1.2.1",
    "qdrant-client==1.15.1",
    "requests==2.32.3",
    "ruff>=0.14.4",
    "sqlalchemy==2.0.44",
    "uvicorn[standard]==0.38.0",
]
